pyproject.toml
[project]
name = "mcp-create-timeseries-chart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.5.0",
    "fastmcp>=2.2.7",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "statsmodels>=0.14.4",
    "vl-convert-python>=1.7.0",
    "yfinance>=0.2.58",
]
